Move to a function and remove the "index" column in the disturbance_matrix_value table

+++ scripts/conversion/update_aidb_1x_to_2x.py
@@ -1,7 +1,26 @@
 """Update the AIDB from libcbm version 1 to libcbm version 2
 
-Use path to the AIDB repo defined in
-So that we modify those files directly
+Use the path to the AIDB repo defined in `eu_cbm_aidb_dir` so that we modify
+those files directly.
+
+
+Usage convert the AIDB from V1 to V2 in all countries:
+
+    # Checkout branch "2.x" of cbm_defaults
+    cd ~/eu_cbm/cbm_defaults/
+    git checkout 2.x
+    # Convert the AIDBs from version 2 to version 2
+    ipython -i ~/eu_cbm/eu_cbm_hat/scripts/conversion/update_aidb_1x_to_2x.py
+
+Usage on one country only for debugging purposes to reproduce errors in a
+specific country:
+
+    >>> import sys
+    >>> import pathlib
+    >>> sys.path.append(str(pathlib.Path.home() /  "eu_cbm/eu_cbm_hat/scripts/conversion/"))
+    >>> from update_aidb_1x_to_2x import convert_aidb_to_v2
+    >>> from update_aidb_1x_to_2x import countries_dir
+    >>> convert_aidb_to_v2(countries_dir / "SI/aidb.db")
 
 See also:
 
@@ -12,6 +31,7 @@ See also:
     - An issue encountered when trying to use libcbm 2
     https://github.com/cat-cfs/libcbm_py/issues/58
 
+
 """
 
 import pathlib
@@ -20,33 +40,67 @@ import pandas
 from cbm_defaults.update import db_updater
 from eu_cbm_hat import eu_cbm_aidb_dir
 
-countries_dir = pathlib.Path(eu_cbm_aidb_dir) / "countries"
-aidbs = countries_dir.glob("**/aidb.db")
 
-for db_path in aidbs:
-    # db_path = countries_dir / "ZZ" / "aidb.db"
-    assert db_path.exists()
+from plumbing.databases.sqlite_database import SQLiteDatabase
+
+
+countries_dir = pathlib.Path(eu_cbm_aidb_dir) / "countries"
+
+
+def convert_aidb_to_v2(db_path):
+    """Convert the aidb at the given path from V1 to V2"""
+    if not db_path.exists():
+        raise ValueError(f"There is no AIDB at {db_path}")
     # If it's already a version 2 table, do nothing
     df = pandas.read_sql_table("land_class", "sqlite:///" + str(db_path))
     if "land_type_id_1" in df.columns:
-        msg = f"\n\n{db_path} is already a version 2 table because the land_class table "
-        msg += "contains the land_type_id_1 column."
+        msg = f"\n\n{db_path} is already a version 2 table because "
+        msg += "the land_class table contains the land_type_id_1 column. "
         msg += "skip conversion."
         print(msg)
-        continue
+        return
     db_path_v1 = db_path.parent / "aidb_v1.db"
     db_path_v2 = db_path.parent / "aidb_v2.db"
-    # Rename the old table to v1
+    # Rename the old AIDB to v1
     shutil.copy(db_path, db_path_v1)
-    # Remove V2 if present
+    # Remove the "index" column in the disturbance_matrix_value table.
+    # This was introduced by mistake and shouldn't be there.
+    db_v1 = SQLiteDatabase(str(db_path_v1))
+    dm_values = db_v1.read_df("disturbance_matrix_value")
+    dm_values = pandas.read_sql_table(
+        "disturbance_matrix_value", "sqlite:///" + str(db_path_v1)
+    )
+    if "index" in dm_values.columns:
+        dm_values.drop(columns="index", inplace=True)
+        dm_values.to_sql(
+            "disturbance_matrix_value",
+            con="sqlite:///" + str(db_path_v1),
+            index=False,
+            if_exists="replace",
+        )
+    # Remove the V2 DB if already present
     if db_path_v2.exists():
         db_path_v2.unlink()
-    try:
-        # Update the AIDB from V1 to V2
-        db_updater.update("1x_to_2x", db_path_v1, db_path_v2)
-        # Make v2 the main table
-        shutil.copy(db_path_v2, db_path)
-        print(f"\n\n{db_path} updated to V2")
-    except Exception as e:
-        print(f"\n\nError {db_path}")
-        print(e)
+    # Update the AIDB from V1 to V2
+    db_updater.update("1x_to_2x", db_path_v1, db_path_v2)
+    # Make v2 the main table
+    shutil.copy(db_path_v2, db_path)
+    print(f"\n{db_path} updated to V2\n\n")
+
+
+if __name__ == "__main__":
+    aidbs = countries_dir.glob("**/aidb.db")
+    for this_db in aidbs:
+        # this_db = countries_dir / "ZZ" / "aidb.db"
+        try:
+            convert_aidb_to_v2(this_db)
+        # Catch all errors. This can be restricted to fewer errors.
+        # Generally observed errors are:
+        # - sqlite3.OperationalError table disturbance_matrix_value has no
+        #   column named index
+        # - sqlite3.IntegrityError UNIQUE constraint failed:
+        #   vol_to_bio_factor.id
+        except Exception as e:
+            print(f"Error {this_db}")
+            print(e)
+            print("\n\n")
